add cake controller and cake view. we can show list of cake and detail of one

+++ src/Controller/CakeController.php
@@ -2,14 +2,19 @@
 
 namespace App\Controller;
 
+use _PHPStan_0a43b4828\Nette\Neon\Exception;
 use App\Entity\Cake;
 use App\Repository\CakeRepository;
 use App\Form\CakeType;
+use phpDocumentor\Reflection\Types\String_;
+use phpDocumentor\Reflection\Types\Void_;
 use Sensio\Bundle\FrameworkExtraBundle\Configuration\Entity;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\Routing\Annotation\Route;
+
+use function PHPUnit\Framework\throwException;
 
 #[Route('/cake')]
 class CakeController extends AbstractController
@@ -72,9 +77,11 @@ class CakeController extends AbstractController
     #[Route('/{id}', name: 'app_cake_delete', methods: ['POST'])]
     public function delete(Request $request, Cake $cake, CakeRepository $cakeRepository): Response
     {
+        if ($request != String_::class) {
+            throw new Exception('Impossible de supprimer le gateau');
+        }
         if ($this->isCsrfTokenValid('delete' . $cake->getId(), $request->request->get('_token'))) {
             $cakeRepository->remove($cake, true);
-            return $this->redirectToRoute('app_cake_index', [], Response::HTTP_SEE_OTHER);
         }
 
         return $this->redirectToRoute('app_cake_index', [], Response::HTTP_SEE_OTHER);
